pyproject.toml
[project]
name = "ai-startup"
version = "0.1.0"
description = "An autonomous system that continuously improves its own codebase using AI agents"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    # Web framework
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    
    # AI and APIs
    "google-generativeai>=0.3.2",
    "PyGithub>=2.1.1",
    
    # Database and ORM
    "sqlalchemy==2.0.23",
    "pydantic==2.5.2",
    "pydantic-settings>=2.1.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.41",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    
    # Code quality
    "black>=24.1.1",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
asyncio_default_fixture_loop_scope = "function"
asyncio_default_test_loop_scope = "function" 